+++ bin/cli.js
@@ -46,39 +46,30 @@ function runTransform({ files, flags, transformer }) {
 }
 
 const TRANSFORMER_INQUIRER_CHOICES = [
-  {
-    name: 'Divider',
-    value: 'Divider'
-  },
-  {
-    name: 'Breadcrumb',
-    value: 'Breadcrumb'
-  },
-  {
-    name: 'Notification',
-    value: 'Notification'
-  },
-  {
-    name: 'Table',
-    value: 'Table'
-  },
-  {
-    name: 'Tabs',
-    value: 'Tabs'
-  },
-  {
-    name: 'Message',
-    value: 'Message'
-  },
-  {
-    name: 'Tooltip',
-    value: 'Tooltip'
-  },
-  {
-    name: 'Modal',
-    value: 'Modal'
-  },
+  'Divider',
+  'Breadcrumb',
+  'Notification',
+  'Table',
+  'Tabs',
+  'Message',
+  'Tooltip',
+  'Modal',
+  'Drawer',
+  'Spin',
+  'Empty',
+  'Slider',
+  'Popover',
+  'Col',
+  'Row',
+  'Tag',
+  'Timeline',
+  'Collapse'
 ]
+  .sort((a, b) => a.localeCompare(b))
+  .map((v) => ({
+    name: v,
+    value: v
+  }))
 
 function expandFilePathsIfNeeded(filesBeforeExpansion) {
   const shouldExpandFiles = filesBeforeExpansion.some((file) =>